Added namespace, composer definition, and tests. Fixed new line slug behavior.
