+++ README
@@ -1,29 +1,27 @@
-= StringToSlug
+# StringToSlug
 
 StringToSlug is a simple class that generate slug versions of a given string, whith
 the capability to swap accented (latin) characters to the "ascii" version.
 
-== License 
+## Examples
 
-(The MIT License)
+```php
+    //Will print text-in-parenthesis-text-in-square-brackets-text-in-curly-brackets
+    echo StringToSlug::gen("(Text in parenthesis) [Text in square brackets] {Text in curly brackets}").PHP_EOL;
 
-Copyright (c) 2008 - 2010 Samir Braga <samirbraga@gmail.com>
+    //Will print a-e-i-o-u-a-e-i-o-u-a-e-i-o-u-a-e-i-o-u-a-o-n-c
+    echo StringToSlug::gen("á é í ó ú à è ì ò ù ä ë ï ö ü â ê î ô û ã õ ñ ç"").PHP_EOL;
 
-Permission is hereby granted, free of charge, to any person obtaining
-a copy of this software and associated documentation files (the
-'Software'), to deal in the Software without restriction, including
-without limitation the rights to use, copy, modify, merge, publish,
-distribute, sublicense, and/or sell copies of the Software, and to
-permit persons to whom the Software is furnished to do so, subject to
-the following conditions:
+    //Will print a-e-i-o-u-a-e-i-o-u-a-e-i-o-u-a-e-i-o-u-a-o-n-c
+    echo StringToSlug::gen("Á É Í Ó Ú À È Ì Ò Ù Ä Ë Ï Ö Ü Â Ê Î Ô Û Ã Õ Ñ Ç").PHP_EOL;
 
-The above copyright notice and this permission notice shall be
-included in all copies or substantial portions of the Software.
+    //Will print symbols-and-punctuations
+    echo StringToSlug::gen("Symbols: @ #  % ^ & * ~ | \ /, and Punctuations ! ?").PHP_EOL;
 
-THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+    //Will print testing-how-a-string-will-resultin-slug-if-have-new-linescharacters
+    echo StringToSlug::gen("Testing how a string \n \n\nwill result\rin slug\r\nif have new lines\rcharacters").PHP_EOL;
+
+    //Will print testing-how-a-string-will-resultin-slugif-have-new-linescharacters
+    StringToSlug::replace_new_lines(false);
+    echo StringToSlug::gen("Testing how a string \n \n\nwill result\rin slug\r\nif have new lines\rcharacters").PHP_EOL;
+```
